refac: course and teacher report functions moved to managers
