+++ managers/jquery_datatable/jquery_datatable_lib.php
@@ -1,5 +1,5 @@
 <?php
-namespace datatable;
+namespace jquery_datatable;
 
 
 /**
