+++ managers/student_profile/studentprofile_lib.php
@@ -123,6 +123,23 @@ function update_user_image_profile($mdl_user_id, $php_file) {
  * @return object --> with CONDICIÓN DE EXCEPCIÓN information
  */
 
+
+/**
+ * Return a moodle url for student profile given a student code by input
+ * @param $courseid
+ * @param $instanceid
+ * @param $student_code
+ * @return moodle_url
+ */
+function get_student_profile_url($courseid, $instanceid, $student_code): moodle_url {
+    return new moodle_url('/blocks/ases/student_profile.php',
+        array(
+            'courseid' => $courseid,
+            'instanceid'=> $instanceid,
+            'student_code' => $student_code
+            ));
+}
+
 function  get_cond_excepcion()
 {
     global $DB;
